Create model module in experiment
